Refactor data fetching into DataUpdateCoordinator
