+++ custom_components/zwift/button.py
@@ -15,14 +15,7 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up Zwift buttons from a config entry."""
-    zwift_data = hass.data[DOMAIN][entry.entry_id]
-
-    entities = []
-    for player_id in zwift_data.players:
-        player = zwift_data.players[player_id]
-        entities.append(ZwiftUpdateButton(player, zwift_data, entry))
-
-    async_add_entities(entities)
+    async_add_entities(hass.data[DOMAIN][entry.entry_id]["entities"]["button"])
 
 
 class ZwiftUpdateButton(ButtonEntity):
@@ -31,9 +24,9 @@ class ZwiftUpdateButton(ButtonEntity):
     _attr_has_entity_name = True
     _attr_translation_key = "update"
 
-    def __init__(self, player, zwift_data, entry):
+    def __init__(self, player, coordinator, entry):
         self._player = player
-        self._zwift_data = zwift_data
+        self._coordinator = coordinator
         self._entry = entry
         self._attr_unique_id = f"zwift_update_{player.player_id}"
 
@@ -47,6 +40,4 @@ class ZwiftUpdateButton(ButtonEntity):
 
     async def async_press(self) -> None:
         """Trigger a manual update for this player."""
-        await self.hass.async_add_executor_job(
-            self._zwift_data.update_player, self._player.player_id
-        )
+        await self._coordinator.async_request_refresh()
